Updated /api/shop
Coded a POST method into api/shop to allow users to purchase items from the shop.
The code sets prices for each item, validates the items, reads the user file and checks if the user is logged in.
Uses the helpers.ts function canAfford to check if the user can afford the item. Deducts the cost accordingly and updates users.json with the new info