Updated /api/auth/register
Updated the backend of the registration page with a POST method to enable the user to register themselves as a user.
Requested info like the name and password from the user then checks if this user exists or not. Hashed the password using bcrypt and then gave the user a unique ID by checking the last user ID and incrementing by 1.
Created the new user object then pushed the data to the users array. Updated users.json to hold the new user