updated actions backend
Added a POST method to actions which included feed, play and return. Read existing pet and user data and checked to see whether the user is logged in and if the pet belongs to them or not. Implemented a switch to check the actions and used functions from helpers.ts to check affordability of the user. Updated budget, pet hunger and happiness accordingly. Wrote the new info into the files.